+++ src/static_versions/v7.rs
@@ -4,7 +4,7 @@
 //! - <https://www.cisco.com/en/US/technologies/tk648/tk362/technologies_white_paper09186a00800a3db9.html>
 
 use crate::protocol::ProtocolTypes;
-use crate::{NetflowPacket, NetflowParseError, ParsedNetflow};
+use crate::{NetflowPacket, NetflowParseError, ParsedNetflow, PartialParse};
 
 use nom::number::complete::be_u32;
 use nom_derive::*;
@@ -16,7 +16,13 @@ use std::net::Ipv4Addr;
 pub(crate) fn parse_netflow_v7(packet: &[u8]) -> Result<ParsedNetflow, NetflowParseError> {
     V7::parse(packet)
         .map(|(remaining, v7)| ParsedNetflow::new(remaining, NetflowPacket::V7(v7)))
-        .map_err(|e| NetflowParseError::V7(e.to_string()))
+        .map_err(|e| {
+            NetflowParseError::Partial(PartialParse {
+                version: 7,
+                error: e.to_string(),
+                remaining: packet.to_vec(),
+            })
+        })
 }
 
 #[derive(Debug, Nom, Clone, Serialize)]